+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Allow special characters in tags. Also Support space-delimited tags:
+  "one two three" and "one, two, three" are equivalent
 
 ## [4.10.0] - 2022-01-19
 ### Added